requirements.txt
appdirs>=1.4.0
Cython>=0.25.2
numpy>=1.12.0
opencv-python>=3.2.0.6
packaging>=16.8
protobuf>=3.2.0
pyparsing>=2.1.10
six>=1.10.0
tensorflow-gpu>=1.0.0
